Factor out `model`
