+++ async-openai/src/types/mod.rs
@@ -5,6 +5,7 @@ mod assistant_file;
 mod assistant_impls;
 mod message;
 mod message_file;
+mod model;
 mod run;
 mod step;
 mod thread;
@@ -13,6 +14,7 @@ pub use assistant::*;
 pub use assistant_file::*;
 pub use message::*;
 pub use message_file::*;
+pub use model::*;
 pub use run::*;
 pub use step::*;
 pub use thread::*;
