+++ async-openai/src/types/types.rs
@@ -3,29 +3,10 @@ use std::{collections::HashMap, path::PathBuf, pin::Pin};
 use bytes::Bytes;
 use derive_builder::Builder;
 use futures::Stream;
-use reqwest::{Body};
+use reqwest::Body;
 use serde::{Deserialize, Serialize};
 
 use crate::error::OpenAIError;
-
-/// Describes an OpenAI model offering that can be used with the API.
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq)]
-pub struct Model {
-    /// The model identifier, which can be referenced in the API endpoints.
-    pub id: String,
-    /// The object type, which is always "model".
-    pub object: String,
-    /// The Unix timestamp (in seconds) when the model was created.
-    pub created: u32,
-    /// The organization that owns the model.
-    pub owned_by: String,
-}
-
-#[derive(Debug, Deserialize, Clone, PartialEq, Serialize)]
-pub struct ListModelResponse {
-    pub object: String,
-    pub data: Vec<Model>,
-}
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
 #[serde(untagged)]
